Fix NPE on DefaultVirtualPacketProvider deactivate

+++ incubator/net/src/main/java/org/onosproject/incubator/net/virtual/impl/provider/DefaultVirtualPacketProvider.java
@@ -110,7 +110,9 @@ public class DefaultVirtualPacketProvider extends AbstractVirtualProvider
 
     @Deactivate
     public void deactivate() {
-        packetService.removeProcessor(processor);
+        if (processor != null) {
+            packetService.removeProcessor(processor);
+        }
     }
 
     @Modified
